requirements.txt
fastapi==0.128.0
pydantic==2.12.5
requests==2.32.5
uvicorn==0.40.0
python-multipart==0.0.9
google-cloud-vision==3.7.1
python-dotenv==1.0.1
aiofiles==24.1.0
pytest==8.3.3
pytest-asyncio==0.24.0
gspread==6.1.0
google-auth==2.37.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
